refactor(infrastructure): remove deprecated shim modules
Remove backward-compatibility shim modules that were only providing
re-exports from canonical locations:

- infrastructure/config/models.py → domain/configs
- infrastructure/files/csv_record_source.py → application/files/csv_record_source

Update all test imports to use direct imports from the canonical modules.
